Separate secret convert code into its own file
